% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aberrance-package.R
\docType{package}
\name{aberrance-package}
\alias{aberrance}
\alias{aberrance-package}
\title{aberrance: Detect Aberrant Behavior in Test Data}
\description{
Detect several types of aberrant behavior, including answer copying, answer similarity, nonparametric misfit, parametric misfit, preknowledge, rapid guessing, and test tampering.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kyliegorney/aberrance}
  \item Report bugs at \url{https://github.com/kyliegorney/aberrance/issues}
}

}
\author{
\strong{Maintainer}: Kylie Gorney \email{kgorney@msu.edu} (\href{https://orcid.org/0000-0002-8924-0726}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Jiayi Deng (\href{https://orcid.org/0000-0002-1962-2956}{ORCID}) [copyright holder]
}

}
\keyword{internal}
